add neaded classes and databases
