+++ Nadim.CinemaReservationSystem.Database/Nadim.CinemaReservationSystem.Database.sqlproj
@@ -64,5 +64,11 @@
   </ItemGroup>
   <ItemGroup>
     <Build Include="dbo\Tables\Users.sql" />
+    <Build Include="dbo\Tables\Cinemas.sql" />
+    <Build Include="dbo\Tables\CinemaRooms.sql" />
+    <Build Include="dbo\Tables\Seats.sql" />
+  </ItemGroup>
+  <ItemGroup>
+    <RefactorLog Include="Nadim.CinemaReservationSystem.Database.refactorlog" />
   </ItemGroup>
 </Project>
