+++ Nadim.CinemaReservationSystem.Database/dbo/Tables/Cinemas.sql
@@ -1,0 +1,6 @@
+﻿CREATE TABLE [dbo].[Cinemas]
+(
+	[CinemaId] INT IDENTITY(1,1) NOT NULL PRIMARY KEY, 
+    [City] NVARCHAR(MAX) NOT NULL, 
+    [Name] NVARCHAR(MAX) NOT NULL
+)
