+++ Nadim.CinemaReservationSystem.Web/Models/CinemaReservationSystemContext.cs
@@ -10,5 +10,8 @@ namespace Nadim.CinemaReservationSystem.Web.Models
     {
         public CinemaReservationSystemContext(DbContextOptions<CinemaReservationSystemContext> options): base(options) { }
         public DbSet<User> Users { get; set; }
+        public DbSet<Cinema> Cinemas { get; set; }
+        public DbSet<CinemaRoom> CinemaRooms { get; set; }
+        public DbSet<Seat> Seats { get; set; }
     }
 }
