+++ Nadim.CinemaReservationSystem.Web/Models/DatabaseModels/CinemaRoom.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Nadim.CinemaReservationSystem.Web.Models
+{
+    public class CinemaRoom
+    {
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int CinemaRoomId { get; set; }
+        public int Number { get; set; }
+
+        public int CinemaId { get; set; }
+        public Cinema Cinema { get; set; }
+
+        public List<Seat> Seats { get; set; }
+    }
+}
