+++ Nadim.CinemaReservationSystem.Web/Models/DatabaseModels/User.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Nadim.CinemaReservationSystem.Web.Models
+{
+    public class User : UserRegistrationInfo
+    {
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int Id { get; set; }
+        public string Role { get; set; }
+    }
+}
