--- Nadim.CinemaReservationSystem.Web/Models/UserLoginInfo.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Nadim.CinemaReservationSystem.Web.Models
-{
-    public class UserLoginInfo
-    {
-        public string Email { get; set; }
-        public string Password { get; set; } 
-    }
-}
