--- Nadim.CinemaReservationSystem.Web/Models/UserRegistrationInfo.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Nadim.CinemaReservationSystem.Web.Models
-{
-    public class UserRegistrationInfo : UserLoginInfo
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-    }
-}
